backend: fix vulnerability that let user update his photo with someone else link photo and when he remove it the photo was removed also from the original user.

+++ backend/photo.py
@@ -9,6 +9,7 @@ from fastapi.responses import FileResponse
 from dotenv import load_dotenv
 import requests
 from auth import check_guest_or_blocked
+from db.utils import is_safe_remove_photo
 
 router = APIRouter(prefix="/photo")
 load_dotenv()
@@ -68,9 +69,11 @@ def uploaded_files_endpoint(photo_name: str) -> UploadFile:
     return FileResponse(file_path)
 
 def delete_photo(photo_link: str) -> None:
-    photo_path = Path(f'{UPLOAD_DIR}/{photo_link.split("/")[-1]}')
+    photo_name = f'{photo_link.split("/")[-1]}'
+    photo_path = Path(f'{UPLOAD_DIR}/{photo_name}')
     try:
-        photo_path.unlink()
+        if is_safe_remove_photo(photo_name):
+            photo_path.unlink()
     except FileNotFoundError:
         logger.warning('Could not delete previous photo %s', photo_path)
